requirements.txt
accelerate==0.30.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.43.1
braceexpand==0.1.7
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
dask==2024.5.1
datasets==2.20.0
decord==0.6.0
# deepspeed==0.14.2
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
editdistance==0.8.1
einops==0.8.0
einops-exts==0.0.4
email-validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.3
filelock==3.14.0
fire==0.6.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
gitpython==3.1.43
gradio==4.39.0
gradio-client==1.1.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
idna==3.7
importlib-resources==6.4.0
iniconfig==2.0.0
jinja2==3.1.4
kiwisolver==1.4.5
locket==1.0.0
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.30.1
opencv-python==4.9.0.80
openpyxl==3.1.4
orjson==3.10.3
pandas==2.2.2
partd==1.4.2
peft==0.11.1
pillow==10.3.0
pluggy==1.5.0
prettytable==3.10.0
protobuf==4.25.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic-core==2.18.2
pydub==0.25.1
pynvml==11.5.0
pyparsing==3.1.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyyaml==6.0.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
ruff==0.5.5
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.2.1
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
smmap==5.0.1
sniffio==1.3.1
spaces==0.29.2
starlette==0.37.2
sympy==1.12
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.40.2
transformers-stream-generator==0.0.5
triton==2.3.0
typer==0.12.3
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
webdataset==0.2.86
websockets==11.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.2